Replace CrudMenuItem with EntityMenuItem
