+++ src/Resources/skeleton/dashboard.tpl
@@ -45,6 +45,6 @@ class <?= $class_name; ?> extends AbstractDashboardController
     public function configureMenuItems(): iterable
     {
         yield MenuItem::linkToDashboard('Dashboard', 'fa fa-home');
-        // yield MenuItem::linkToCrud('The Label', 'fas fa-list', EntityClass::class);
+        // yield MenuItem::linkToEntity(EntityClass::class, 'The Label', 'fas fa-list');
     }
 }
