+++ tests/Functional/Apps/CustomizationApp/src/Controller/Dashboard/RelativeUrlsTestDashboardController.php
@@ -30,6 +30,6 @@ class RelativeUrlsTestDashboardController extends AbstractDashboardController
     public function configureMenuItems(): iterable
     {
         yield MenuItem::linktoDashboard('Dashboard', 'fa fa-home');
-        yield MenuItem::linkToCrud('Demo', 'fas fa-list', DemoEntity::class);
+        yield MenuItem::linkToEntity(DemoEntity::class, 'Demo', 'fas fa-list');
     }
 }
