+++ tests/Functional/Apps/DefaultApp/src/Controller/Dashboard/DisabledDarkModeDashboardController.php
@@ -30,6 +30,6 @@ class DisabledDarkModeDashboardController extends AbstractDashboardController
     public function configureMenuItems(): iterable
     {
         yield MenuItem::linktoDashboard('Dashboard', 'fa fa-home');
-        yield MenuItem::linkToCrud('Categories', 'fas fa-tags', Category::class);
+        yield MenuItem::linkToEntity(Category::class, 'Categories', 'fas fa-tags');
     }
 }
