+++ tests/Functional/Apps/DefaultApp/src/Controller/Dashboard/LayoutDashboardController.php
@@ -38,6 +38,6 @@ class LayoutDashboardController extends AbstractDashboardController
     public function configureMenuItems(): iterable
     {
         yield MenuItem::linktoDashboard('Dashboard', 'fa fa-home');
-        yield MenuItem::linkToCrud('Categories', 'fas fa-tags', Category::class);
+        yield MenuItem::linkToEntity(Category::class, 'Categories', 'fas fa-tags');
     }
 }
