+++ tests/Functional/Apps/DefaultApp/src/Controller/Dashboard/LocaleDashboardController.php
@@ -37,6 +37,6 @@ class LocaleDashboardController extends AbstractDashboardController
     public function configureMenuItems(): iterable
     {
         yield MenuItem::linktoDashboard('Dashboard', 'fa fa-home');
-        yield MenuItem::linkToCrud('Categories', 'fas fa-tags', Category::class);
+        yield MenuItem::linkToEntity(Category::class, 'Categories', 'fas fa-tags');
     }
 }
