+++ tests/Functional/Apps/DefaultApp/src/Controller/Dashboard/UserMenuDashboardController.php
@@ -45,6 +45,6 @@ class UserMenuDashboardController extends AbstractDashboardController
     public function configureMenuItems(): iterable
     {
         yield MenuItem::linktoDashboard('Dashboard', 'fa fa-home');
-        yield MenuItem::linkToCrud('Categories', 'fas fa-tags', Category::class);
+        yield MenuItem::linkToEntity(Category::class, 'Categories', 'fas fa-tags');
     }
 }
