+++ tests/Functional/Apps/SecuredApp/src/Controller/UserMenuDashboardController.php
@@ -45,7 +45,7 @@ class UserMenuDashboardController extends AbstractDashboardController
     public function configureMenuItems(): iterable
     {
         yield MenuItem::linktoDashboard('Dashboard', 'fa fa-home');
-        yield MenuItem::linkToCrud('Categories', 'fas fa-tags', Category::class)
+        yield MenuItem::linkToEntity(Category::class, 'Categories', 'fas fa-tags')
             ->setController(CategoryCrudController::class);
     }
 }
